requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
filelock==3.12.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
Pillow==9.5.0
requests==2.30.0
sqlparse==0.4.4
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
